# backend/docs/swaggerpet.json
{ "openapi":"3.0.0",
  "info":{ 
      "title":"World Happiness Rankings by n9985310",
      "description":"This API has been created to support the assessment in QUT's CAB230 Web Computing unit. It exposes a small number of REST endpoints which implement CRUD operations on a database containing publicly available data from the World Happiness Report [https://en.wikipedia.org/wiki/World_Happiness_Report](https://en.wikipedia.org/wiki/World_Happiness_Report) - an annual survey that ranks the overall happiness of a country according to the views of its citizens on a range of happiness measures. This API provides access to the survey data for the years 2015 to 2020 (inclusive). The API endpoints and their usage are described in detail below.",
      "version":"1.0.0"
  },
  "schemes":[
    "http",
    "https"
  ],
  "basePath":"/v2",
  "tags":[
      { 
        "name":"Data"
      },
      { 
        "name":"Authentication"
      },
      { 
        "name":"Profile"
      }
  ],
  "paths":{ 
      "/rankings":{ 
          "get":{ 
              "tags":["Data"],
              "description":"Returns a list of countries and their happiness rank for the years 2015 to 2020. The list is arranged by year, in descending order. The list can optionally be filtered by year and/or country name using query parameters.",
              "consumes":["application/json","application/xml"],
              "produces":["application/xml","application/json"],
              "parameters":[
                  {
                    "in":"query",
                    "name":"year",
                    "description":"Year of world happiness report",
                    "required":false
                  },
                  {
                    "in":"query",
                    "name":"country",
                    "description":"Name of country",
                    "required":false
                  }
                ],
              "responses":{
                  "200":{
                    "description":"An array of objects containing happiness rank, country, score and year properties. An example of one object in the array is shown below.",
                    "content":{
                      "application/json":{
                        "schema":{
                          "type":"array",
                          "items":{
                            "type":"object",
                            "properties":{
                              "rank":{
                                "type":"integer",
                                "format":"int64",
                                "example":1
                              },
                              "country":{
                                "type":"string",
                                "example":"Finland"
                              },
                              "score":{
                                "type":"string",
                                "example":"7.769"
                              },
                              "year":{
                                "type":"integer",
                                "format":"int64",
                                "example":"2019"
                              }
                            }
                          }
                        }
                      }
                    }
                  },
                  "400":{
                          "description":"Invalid country or year query parameters. Click on 'Schema' below to see the possible error responses.",
                          "content":{
                            "application/json":{
                              "schema":{
                                  "oneOf":{
                                    "$ref":"#/definitions/InvalidCountryFormat",
                                    "$ref":"#/definitions/InvalidYearFormat"
                                  }
                                },
                                "example":{
                                    "$ref":"#/definitions/InvalidCountryFormat"
                                }
                              }
                            }
                          }
                        }
              }
          },
        "/countries":{
          "get":{
            "tags":["Data"],
            "description":"Returns a list of all surveyed countries, ordered alphabetically.",
            "responses":{
              "200":{
                "description":"An array of countries. An example of the first five elements in the array is shown below.",
                "content":{
                  "application/json":{
                    "schema":{
                      "type":"array",
                      "items":{
                        "type":"string"
                      },
                      "example":[  "Afghanistan",
                        "Albania",
                        "Algeria",
                        "Angola",
                        "Argentina"]
                    }
                  }
                }
              },
              "400":{
                "description":"Invalid query parameters.",
                "content":{
                  "application/json":{
                    "schema":{
                      "type":"object",
                      "properties":{
                        "error":{
                          "type":"boolean",
                          "example":true
                        },
                        "message":{
                          "type":"string",
                          "example":"Invalid query parameters. Query parameters are not permitted."
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "/factors/{year}":{
          "get":{
            "tags":["Data"],
            "description":"Returns a list of countries and their associated happiness factor scores for the specified year. The path parameter (year) is required. The number of returned results can be limited by the optional limit query parameter. A result for a single country can be obtained via the optional country query parameter. This route also requires the user to be authenticated - a valid JWT token must be sent in the header of the request. To test this route using Swagger, click the green 'Authorize' button at the top of this page to enter your JWT token. A JWT token can be obtained by logging in.",
            "parameters":[
              {
                "in":"path",
                "name":"year",
                "description":"Year of world happiness report",
                "required":true
              },
              {
                "in":"query",
                "name":"limit",
                "description":"The maximum number of results to return",
                "required":false
              },
              {
                "in":"query",
                "name":"country",
                "description":"Name of country",
                "required":false
              }
            ],
            "responses":{
              "200":{
                "description":"An array of objects containing happiness rank, country, score, economy, family, health, freedom, generosity and trust properties. An example of one object in the array is shown below.",
                "content":{
                  "application/json":{
                    "schema":{
                      "type":"array",
                      "items":{
                        "type":"object",
                        "properties":{
                          "rank":{
                            "type":"integer",
                            "format":"int64",
                            "example":1
                          },
                          "country":{
                            "type":"string",
                            "example":"Finland"
                          },
                          "score":{
                            "type":"string",
                            "example":"7.769"
                          },
                          "economy":{
                            "type":"string",
                            "example":"1.340"
                          },
                          "family":{
                            "type":"string",
                            "example":"1.587"
                          },
                          "health":{
                            "type":"string",
                            "example":"1.340"
                          },
                          "freedom":{
                            "type":"string",
                            "example":"0.596"
                          },
                          "generosity":{
                            "type":"string",
                            "example":"0.153"
                          },
                          "trust":{
                            "type":"string",
                            "example":"0.393"
                          }
                        }
                      }
                    }
                  }
                }
              },
              "400":{
                "description":"Invalid parameters. Click on 'Schema' below to see the possible error responses.",
                "content":{
                  "application/json":{
                    "schema":{
                      "$ref":"#/definitions/YearFormat"
                    }
                  }
                }
              },
              "401":{
                "description":"Unauthorized. Click on 'Schema' below to see the possible error responses.",
                "content":{
                  "application/json":{
                    "schema":{
                      "$ref":"#/definitions/AuthHeaderNotFound"
                    }
                  }
                }
              }
            }
          }
        },
        "/user/register":{
          "post":{
            "tags":["Authentication"],
            "description":"Creates a new user account. A request body containing the user to be registered must be sent.",
            "parameters":[
              {
                "in":"body",
                "name":"body",
                "description":"An object containing the email and password of the user to be registered.",
                "required":true,
                "schema":{
                  "$ref":"#/definitions/UserFormat"
                }
              }
            ],
            "responses":{
              "201":{
                "description":"User successfully created",
                "content":{
                  "application/json":{
                    "schema":{
                      "type":"object",
                      "properties":{
                        "message":{
                          "type":"string",
                          "example":"User created"
                        }
                      }
                    }
                  }
                }
              },
              "400":{
                "description":"Bad request",
                "content":{
                  "application/json":{
                    "schema":{
                      "$ref":"#/definitions/UserBadRequest"
                    }
                  }
                }
              },
              "409":{
                "description":"User already exists",
                "content":{
                  "application/json":{
                    "schema":{
                      "type":"object",
                      "properties":{
                        "error":{
                          "type":"boolean",
                          "example":true
                        },
                        "message":{
                          "type":"string",
                          "example":"User already exists"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "/user/login":{
          "post":{
            "tags":["Authentication"],
            "description":"Log in to an existing user account. A request body containing the user credentials must be sent.",
            "parameters":[
              {
                "in":"body",
                "name":"body",
                "description":"An object containing the email and password of the user to be registered.",
                "required":true,
                "schema":{
                  "$ref":"#/definitions/UserFormat"
                }
              }
            ],
            "responses":{
              "200":{
                "description":"Log in successful",
                "content":{
                  "application/json":{
                    "schema":{
                      "type":"object",
                      "properties":{
                        "token":{
                          "type":"string",
                          "example":"ajsonwebtoken"
                        },
                        "token_type":{
                          "type":"string",
                          "example":"Bearer"
                        },
                        "expires_in":{
                          "type":"integer",
                          "format":"int64",
                          "example":86400
                        }
                      }
                    }
                  }
                }
              },
              "400":{
                "description":"Bad request",
                "content":{
                  "application/json":{
                    "schema":{
                      "$ref":"#/definitions/UserBadRequest"
                    }
                  }
                }
              },
              "401":{
                "description":"Log in failed",
                "content":{
                  "application/json":{
                    "schema":{
                      "type":"object",
                      "properties":{
                        "error":{
                          "type":"boolean",
                          "example":true
                        },
                        "message":{
                          "type":"string",
                          "example":"Incorrect email or password"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "/user/{email}/profile":{
          "get":{
            "tags":["Profile"],
            "description":"Returns an object containing a user's profile information. The path parameter (email) is required and must be an email that has been registered. This route returns additional information if the user is authenticated. To be authenticated, a valid JWT token must be sent in the header of the request. To test this using Swagger, click the green 'Authorize' button at the top of this page to enter your JWT token. A JWT token can be obtained by logging in.",
            "parameters":[
              {
                "in":"path",
                "name":"email",
                "description":"Email address of a registered user",
                "required":true
              }
            ],
            "responses":{
              "200":{
                "description":"An object containing email, firstName, lastName, address (authenticated only) and dob (authenticated only) properties. An example object for an authenticated request is shown below. If no profile information has been entered for the user via the profile PUT request route, all returned values except email are expected to be null. Click on 'Schema' below to see a response example for an unauthenticated request.",
                "content":{
                  "application/json":{
                    "schema":{
                      "$ref":"#/definitions/EmailFormat"
                    }
                  }
                }
              },
              "401":{
                "description":"Unauthorized. Click on 'Schema' below to see the possible error responses.",
                "content":{
                  "application/json":{
                    "schema":{
                      "type":"object",
                      "properties":{
                        "error":{
                          "type":"boolean",
                          "example":true
                        },
                        "message":{
                          "type":"string",
                          "example":"JWT token has expired"
                        }
                      }
                    }
                  }
                }
              },
              "409":{
                "description":"User not found",
                "content":{
                  "application/json":{
                    "schema":{
                      "type":"object",
                      "properties":{
                        "error":{
                          "type":"boolean",
                          "example":true
                        },
                        "message":{
                          "type":"string",
                          "example":"User not found"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "put":{
            "tags":["Profile"],
            "description":"Updates a user's profile information. The path parameter (email) is required and must be an email that has been registered. A request body containing the user's profile information must be sent. This route returns an object containing the user's updated profile information. This route also requires the user to be authenticated - a valid JWT token must be sent in the header of the request. To test this route using Swagger, click the green 'Authorize' button at the top of this page to enter your JWT token. A JWT token can be obtained by logging in.",
            "parameters":[
              {
                "in":"path",
                "name":"email",
                "description":"Email address of a registered user",
                "required":true
              },
              {
                "in":"body",
                "name":"body",
                "required":false,
                "schema":{
                  "type":"object",
                  "properties":{
                    "firstName":{
                      "type":"string",
                      "example":"Michael"
                    },
                    "lastName":{
                      "type":"string",
                      "example":"Jordan"
                    },
                    "dob":{
                      "type":"string",
                      "example":"1963-02-17"
                    },
                    "address":{
                      "type":"string",
                      "example":"123 Fake Street, Springfield"
                    }
                  }
                }
              }
            ],
            "responses":{
              "201":{
                "description":"User successfully created",
                "content":{
                  "application/json":{
                    "schema":{
                      "$ref":"#/definitions/EmailFormat"
                    }
                  }
                }
              },
              "400":{
                "description":"Bad request. Click on 'Schema' below to see the possible error responses",
                "content":{
                  "application/json":{
                    "schema":{
                      "type":"object",
                      "properties":{
                        "error":{
                          "type":"boolean",
                          "example":true
                        },
                        "message":{
                          "type":"string",
                          "example":"Request body incomplete: firstName, lastName, dob and address are required."
                        }
                      }
                    }
                  }
                }
              },
              "401":{
                "description":"Unauthorized. Click on 'Schema' below to see the possible error responses.",
                "content":{
                  "application/json":{
                    "schema":{
                      "$ref":"#/definitions/AuthHeaderNotFound"
                    }
                  }
                }
              },
              "403":{
                "description":"Forbidden. Email address associated with JWT token is not the same as email provided in path parameter.",
                "content":{
                  "application/json":{
                    "schema":{
                      "type":"object",
                      "properties":{
                        "error":{
                          "type":"boolean",
                          "example":true
                        },
                        "message":{
                          "type":"string",
                          "example":"Forbidden"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
    "securityDefinitions":{
      "Bearer":{
        "type":"apikey",
        "name":"Authorization",
        "in":"header"
      },
      "petstore_auth":{
        "type": "oauth2",
        "authorizationUrl": "http://petstore.swagger.io/oauth/dialog",
        "flow": "implicit",
        "scopes":{
          "write":{
            "pets": "modify pets in your account"
          },
          "read":{
            "pets": "read your pets"
          }
        }
      },
      "api_key":{
        "type": "apiKey",
        "name": "api_key",
        "in": "header"
      }
    },
    "definitions":{
      "YearFormat":{
        "type":"object",
        "properties":{
          "error":{
            "type":"boolean",
            "example":true
          },
          "message":{
            "type":"string",
            "example":"Invalid year format. Format must be yyyy."
          }
      }
    },
    "AuthHeaderNotFound":{
      "type":"object",
      "properties":{
        "error":{
          "type":"boolean",
          "example":true
        },
        "message":{
          "type":"string",
          "example":"Authorization header ('Bearer token') not found"
        }
      }
    },
    "InvalidCountryFormat":{
      "type":"object",
      "properties":{
        "error":{
          "type":"boolean",
          "example":true
        },
        "message":{
          "type":"string",
          "example":"Invalid country format. Country query parameter cannot contain numbers."
        }
      }
    },
    "InvalidYearFormat":{
      "type":"object",
      "properties":{
        "error":{
          "type":"boolean",
          "example":true
        },
        "message":{
          "type":"string",
          "example":"Invalid year format. Format must be yyyy."
        }
      }
    },
    "UserFormat":{
      "type":"object",
      "properties":{
        "email":{
          "type":"string",
          "example":"mike@gmail.com"
        },
        "password":{
          "type":"string",
          "example":"passport"
        }
      }
    },
    "UserBadRequest":{
      "type":"object",
      "properties":{
        "error":{
          "type":"boolean",
          "example":true
        },
        "message":{
          "type":"string",
          "example":"Request body incomplete, both email and password are required"
        }
      }
    },
    "EmailFormat":{
      "type":"object",
      "properties":{
        "email":{
          "type":"string",
          "example":"mike@gmail.com"
        },
        "firstName":{
          "type":"string",
          "example":"Michael"
        },
        "lastName":{
          "type":"string",
          "example":"Jordan"
        },
        "dob":{
          "type":"string",
          "example":"1963-02-17"
        },
        "address":{
          "type":"string",
          "example":"123 Fake Street, Springfield"
        }
      }
    }
  }
}
  